refactor: remove provider-level OpenResponses text client, use protocol
OpenResponses is a protocol, not a provider. The provider-level
`modalities/text/providers/openresponses/` was a parallel hierarchy
that reimplemented tool call parsing inline instead of delegating to
the canonical `parse_tool_calls()` from the protocol layer. The inline
versions were also less robust (KeyError vs .get() with fallback).

OpenAI, xAI, and Ollama now inherit from the protocol-level
`OpenResponsesTextClient`, matching the ChatCompletions pattern where
DeepSeek, Groq, Mistral etc. inherit from `ChatCompletionsTextClient`.

Fixes #219
